Fix Bug in Fetching Start/End Dates
Some bug related to fetching the correct academic start date and end
date is resolved. A new HTML web client and parser is used to fix the
problem, but it is still unclear why the old web client seems to fetch
a slightly different HTML string from the registrar page containing a
different school start date and end date.